moved public QoS type to Sample.rs
